+++ tests/BaseBlend.php
@@ -31,8 +31,8 @@ class BaseBlend extends TestCase
     public static function setUpBeforeClass()
     {
         // copy from database.. to temp/database/...
-        $fileHelper = new SimpleCache(BLEND_MODX_MIGRATION_PATH.'temp/');
-        $fileHelper->copyDirectory(BLEND_MODX_MIGRATION_PATH.'database/', BLEND_MODX_MIGRATION_PATH.'temp/database/');
+        $fileHelper = new SimpleCache(BLEND_TEST_MIGRATION_PATH.'temp/');
+        $fileHelper->copyDirectory(BLEND_TEST_MIGRATION_PATH.'database/', BLEND_TEST_MIGRATION_PATH.'temp/database/');
     }
 
     /**
@@ -42,8 +42,8 @@ class BaseBlend extends TestCase
     {
         if (BLEND_CLEAN_UP) {
             // delete files form temp:
-            $fileHelper = new SimpleCache(BLEND_MODX_MIGRATION_PATH.'temp/');
-            $fileHelper->deleteDirectory(BLEND_MODX_MIGRATION_PATH.'temp/database/');
+            $fileHelper = new SimpleCache(BLEND_TEST_MIGRATION_PATH.'temp/');
+            $fileHelper->deleteDirectory(BLEND_TEST_MIGRATION_PATH.'temp/database/');
         }
     }
 
@@ -80,7 +80,7 @@ class BaseBlend extends TestCase
 
         $this->consoleUserInteractionHandler = new VoidUserInteractionHandler();
 
-        $this->blender = new Blender($this->modx, $this->consoleUserInteractionHandler, ['blend_modx_migration_dir' => BLEND_MODX_MIGRATION_PATH.'temp/']);
+        $this->blender = new Blender($this->modx, $this->consoleUserInteractionHandler, ['blend_modx_migration_dir' => BLEND_TEST_MIGRATION_PATH.'temp/']);
 
         if ($this->install_blend) {
             $this->blender->install();